Create Task 1: TO-DO LIST
This project aims to create a GUI-based application using Python, allowing users to create, update and track their to-do lists i.e. the task is In-Progress or Completed or Pending. In addition to this Delete button is also added from which any existing task irrespective of its status can be deleted by user.